bootstrap/templates/default/requirements.txt
bootstrap.pytorch